requirements.txt
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
astroid==3.1.0
Authlib==1.6.6
bcrypt==5.0.0
beautifulsoup4==4.14.3
boto3==1.40.61
botocore==1.40.61
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
coverage==7.13.1
coze-coding-dev-sdk==0.5.4
coze-coding-utils==0.1.11
coze-workload-identity==0.1.4
cozeloop==0.1.21
cryptography==46.0.3
cssselect==1.3.0
dbus-python==1.3.2
dill==0.4.0
distro==1.9.0
docx2python==3.5.0
et_xmlfile==2.0.0
fastapi==0.121.2
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.8.2
idna==3.11
inflect==7.5.0
iniconfig==2.3.0
isort==5.13.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.3
langchain-core==1.0.2
langchain-openai==1.0.1
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-checkpoint-postgres==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.39
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.8.0
numpy==2.2.6
openai==2.14.0
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.2.2
paragraphs==1.0.1
pillow==10.3.0
platformdirs==4.5.1
pluggy==1.6.0
psutil==7.1.3
psycopg==3.3.0
psycopg-binary==3.3.0
psycopg-pool==3.3.0
psycopg2-binary==2.9.9
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyGObject==3.48.2
PyJWT==2.10.1
pylint==3.1.0
pypdf==6.4.1
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
s3transfer==0.14.0
setuptools==68.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
sqlacodegen==3.2.0
SQLAlchemy==2.0.44
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tomlkit==0.13.3
tqdm==4.67.1
typeguard==4.4.4
types-html5lib==1.1.11.20251117
types-lxml==2026.1.1
types-webencodings==0.5.0.20251108
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.38.0
watchdog==6.0.0
websockets==15.0.1
wheel==0.42.0
wsproto==1.3.2
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
zstandard==0.25.0
